Fixed bug: scrolling to the bottom of the timeline
The method that scrolls to the bottom of the current timeline:
`back_on_bottom` in `interface.py` raised an AttributeError because it
was accesing a nonexistent attribute.

I added a `status_count` method to the Timeline class in `timeline.py`
which updates and returns the count of the tweets, and I have used
the value that it returns to set the focus at the bottom of the
timeline.